requirements.txt
pdf2docx==0.5.6
langchain~=0.0.288
tiktoken
python-lorem